agent_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-adk
google-genai
ag-ui-adk
composio
httpx
PyJWT
